fix(flamegraph): move flamegraph template code to own file and add license
