+++ .cargo/config.toml
@@ -1,7 +1,7 @@
 [alias]
 ci-fmt = "fmt --check"
 ci-clippy = "clippy --all-targets --all-features -- -D warnings"
-ci-license = "license-template --template .license_template"
+ci-license = "license-template --template .license_template --ignore .license_ignore"
 ci-udeps = "udeps --all-targets --backend=depinfo"
 ci-udeps-external = "udeps --all-targets --manifest-path external-crates/move/Cargo.toml --backend=depinfo"
