+++ .license_ignore
@@ -1,0 +1,2 @@
+# Files with non-Apache-2.0 licenses that should be excluded from license template check
+crates/telemetry-subscribers/src/flamegraph/svg_template.rs
